Refactoring de l'authentification avec l'ajouter du domaine d'Healthcheck (avec seulement le PING pour l'instant).
